# swagger.yml
consumes:
  - application/json
definitions:
  error_out:
    type: object
    description: "The standard format for an error response body."
    required:
      - reason
    properties:
      reason:
        type: string
        description: "The reason for the error."
        minLength: 1
...
  resource_update:
    type: object
    description: "A modification to a resource."
    required:
      - name
    properties:
      name:
        type: string
        description: "The new resource name."
        minLength: 1
  resource_out:
    type: object
    description: "An outgoing resource."
    required:
      - id
      - name
      - resource_type
    properties:
      id:
        type: string
        description: "The resource identifier."
        minLength: 36
        maxLength: 36
      name:
        type: string
        description: "The resource name."
        minLength: 1
      resource_type:
        type: string
        description: "The resource type name."
        minLength: 1
...
info:
  description: >-
    Manages Permissions for the CyVerse Discovery Environment and related applications.
  title: "Permissions Service"
  version: "5.2.8.0"
paths:
...
  /resources/{id}:
    parameters:
      - name: id
        type: string
        description: "The resource ID."
        in: path
        required: True
    put:
      tags:
        - resources
      summary: "Update a Resource"
      description: "Updates a resource in the database."
      parameters:
        - description: "The updated resource information."
          in: body
          name: "resource_update"
          required: True
          schema:
            $ref: "#/definitions/resource_update"
      operationId: updateResource
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/resource_out"
        400:
          $ref: "#/responses/bad_request"
        404:
          $ref: "#/responses/not_found"
        500:
          $ref: "#/responses/internal_server_error"
    delete:
      tags:
        - resources
      summary: "Delete a Resource"
      description: "Removes a resource from the database."
      operationId: deleteResource
      responses:
        200:
          description: "OK"
        404:
          $ref: "#/responses/not_found"
        500:
          $ref: "#/responses/internal_server_error"
...
produces:
  - application/json
responses:
  bad_request:
    description: "Bad Request"
    schema:
      $ref: "#/definitions/error_out"
  internal_server_error:
    description: "Internal Server Error"
    schema:
      $ref: "#/definitions/error_out"
  not_found:
    description: "Not Found"
    schema:
      $ref: "#/definitions/error_out"
schemes:
  - http
swagger: "2.0"